backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
imap-tools
google-generativeai
pandas
openpyxl
matplotlib
jinja2
requests
python-dotenv
